moving away from generic read API
also changed heroes bitmask to use identifiers instead

+++ MapObjects/DwellingObject.cs
@@ -8,6 +8,6 @@ namespace H3Mapper.MapObjects
         public Factions AllowedFactions { get; set; }
         public UnitLevel MinLevel { get; set; }
         public UnitLevel MaxLevel { get; set; }
-        public uint SameAsCastle { get; set; }
+        public long SameAsCastle { get; set; }
     }
 }
